label: update create user function return
