+++ src/modules/user/services/admin-user.service.ts
@@ -4,7 +4,7 @@ import { IUserRepository } from '../interfaces/repository.interface';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { UserRole } from '../enum/user-role.enum';
 import { AppError } from '../../../common/errors/Error';
-import { EmployeeInfo } from '@prisma/client';
+import { User } from '../interfaces/user.interface';
 
 @Injectable()
 export class CreateAdminUserService {
@@ -13,7 +13,7 @@ export class CreateAdminUserService {
     private userRepository: IUserRepository,
   ) {}
 
-  async execute(data: CreateUserDto): Promise<EmployeeInfo> {
+  async execute(data: CreateUserDto): Promise<User> {
     const adminSignupToken = process.env.ADMIN_SIGNUP_TOKEN;
     const { signupToken, password, passwordConfirmation } = data;
